refactor: :recycle: 화면 너비 높이 감지하는 로직 훅으로 분리
